requirements.txt
avro-python3==1.10.2
codetiming==1.3.2
confluent-kafka==1.9.2
fastavro==1.4.8
flask==2.2.2
https://github.com/unbit/uwsgi/archive/refs/tags/2.0.21.zip
prometheus-client==0.15.0
requests==2.28.1
